+++ src/CookTime/client-app/src/components/IngredientRequirements/IngredientRequirementList.tsx
@@ -1,18 +1,21 @@
 import React from "react";
 import { Row } from "react-bootstrap";
 import { IngredientRequirement, MeasureUnit } from "src/shared/CookTime";
+import { UnitPreference } from "src/shared/units";
 import { IngredientDisplay } from "../Ingredients/IngredientDisplay";
 
 interface IngredientRequirementListProps {
   ingredientRequirements: IngredientRequirement[];
   units: MeasureUnit[];
   multiplier: number;
+  unitPreference: UnitPreference;
 }
 
 export function IngredientRequirementList({
   ingredientRequirements,
   units,
   multiplier,
+  unitPreference,
 }: IngredientRequirementListProps) {
   return (
     <>
@@ -23,6 +26,7 @@ export function IngredientRequirementList({
             <IngredientDisplay
               showAlternatUnit={true}
               units={units}
+              unitPreference={unitPreference}
               ingredientRequirement={{ ...ingredient, quantity: scaledQuantity }}
             />
           </Row>